Extract class for the leap path
